lesson30/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
scikit-learn==1.5.1
numpy==1.26.4
pillow==10.4.0
joblib==1.4.2
